# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a doc for compass_mini_api.",
        "title": "compass_mini_api",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/employee": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get All Employee",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Get All Employee",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 0,
                        "description": "offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "name / company / division / supervisor / email / joindate",
                        "description": "order",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "asc / desc",
                        "description": "direction",
                        "name": "direction",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "\u003cbr\u003eexample string json: [{\"column\":\"name\",\"value\":\"yusnar\",\"comparation\":\"%\"},{\"column\":\"companyid\",\"value\":\"43\",\"comparation\":\"=\"},{\"column\":\"joindate\",\"value\":\"2024-01-24_2024-01-28\",\"comparation\":\"between\"}]\u003cbr\u003eexample query encode: %5B%7B%22column%22%3A%22name%22%2C%22value%22%3A%22yusnar%22%2C%22comparation%22%3A%22%25%22%7D%2C%7B%22column%22%3A%22companyid%22%2C%22value%22%3A%2243%22%2C%22comparation%22%3A%22%3D%22%7D%2C%7B%22column%22%3A%22joindate%22%2C%22value%22%3A%222024-01-24_2024-01-28%22%2C%22comparation%22%3A%22between%22%7D%5D",
                        "description": "filter by conditions with query encode value",
                        "name": "conditions",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.EmployeeGetAllResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create employee",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Create employee",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "CODE.ID",
                        "name": "company",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 43,
                        "name": "companyid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "Maintenance",
                        "name": "division",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 16,
                        "name": "divisionid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "yusnar@code.id",
                        "name": "email",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "example": "2024-01-24",
                        "name": "joindate",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "Yusnar Setiyadi",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "+6281234567812",
                        "name": "phonenumber",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "example": "Herru Purnomo Santoso",
                        "name": "supervisor",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 11,
                        "name": "supervisorid",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "image",
                        "name": "employeephoto",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/employee/employeephoto/{base64}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Employee Photo",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "multipart/form-data"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Get Employee Photo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "base64",
                        "name": "base64",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/v1/employee/supervisor": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get All Employee for Supervisor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Get All Employee for Supervisor",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 10,
                        "description": "limit",
                        "name": "limit",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 0,
                        "description": "offset",
                        "name": "offset",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\u003cbr\u003eexample string json: [{\"column\":\"name\",\"value\":\"herru\",\"comparation\":\"%\"}]\u003cbr\u003eexample query encode: %5B%7B%22column%22%3A%22name%22%2C%22value%22%3A%22herru%22%2C%22comparation%22%3A%22%25%22%7D%5D",
                        "description": "filter by conditions with query encode value",
                        "name": "conditions",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.EmployeeSupervisorGetAllResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/employee/with_base64": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create employee With Base64, please fill data base64 with Plain-text - just The Base64 value, see https://base64.guru/converter/encode/file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Create employee With Base64",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateEmployeeRequestWithBase64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/employee/with_base64/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update employee With Base64, please fill data base64 with Plain-text - just The Base64 value, see https://base64.guru/converter/encode/file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Update employee With Base64",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "id employee",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "request body",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateEmployeeRequestWithBase64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ResponseMessage"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateEmployeeRequestWithBase64": {
            "type": "object",
            "required": [
                "company",
                "companyid",
                "division",
                "divisionid",
                "employeephoto",
                "joindate",
                "name",
                "supervisor",
                "supervisorid"
            ],
            "properties": {
                "company": {
                    "type": "string",
                    "example": "CODE.ID"
                },
                "companyid": {
                    "type": "integer",
                    "example": 43
                },
                "division": {
                    "type": "string",
                    "example": "Maintenance"
                },
                "divisionid": {
                    "type": "integer",
                    "example": 16
                },
                "email": {
                    "type": "string",
                    "example": "yusnar@code.id"
                },
                "employeephoto": {
                    "$ref": "#/definitions/dto.EmployeePhoto"
                },
                "joindate": {
                    "type": "string",
                    "example": "2024-01-24"
                },
                "name": {
                    "type": "string",
                    "example": "Yusnar Setiyadi"
                },
                "phonenumber": {
                    "type": "string",
                    "example": "+6281234567812"
                },
                "supervisor": {
                    "type": "string",
                    "example": "Herru Purnomo Santoso"
                },
                "supervisorid": {
                    "type": "integer",
                    "example": 11
                }
            }
        },
        "dto.EmployeeGetAllResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "item": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.EmployeeGetAllResponse"
                    }
                }
            }
        },
...
        "dto.EmployeeSupervisorGetAllResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "item": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.EmployeeSupervisorGetAllResponse"
                    }
                }
            }
        },
...
        "dto.ResponseMessage": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "response.errorResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "description": {},
                "meta": {
                    "$ref": "#/definitions/response.Meta"
                }
            }
        }
    }
}